modified:   src/renderers/dom/ReactDOMNodeStreamRenderer.js
