+++ src/renderers/dom/ReactDOMNodeStreamRenderer.js
@@ -14,6 +14,7 @@
 var invariant = require('fbjs/lib/invariant');
 var React = require('react');
 var ReactPartialRenderer = require('ReactPartialRenderer');
+var ReactFeatureFlags = require('ReactFeatureFlags');
 
 var Readable = require('stream').Readable;
 
@@ -49,10 +50,13 @@ function renderToStream(element) {
  * See https://facebook.github.io/react/docs/react-dom-stream.html#rendertostaticstream
  */
 function renderToStaticStream(element) {
-  invariant(
-    React.isValidElement(element),
-    'renderToStaticStream(): You must pass a valid ReactElement.',
-  );
+  const disableNewFiberFeatures = ReactFeatureFlags.disableNewFiberFeatures;
+  if (disableNewFiberFeatures) {
+    invariant(
+      React.isValidElement(element),
+      'renderToString(): You must pass a valid ReactElement.',
+    );
+  }
   return new ReactMarkupReadableStream(element, true);
 }
